# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormAssignReq:
    properties:
      room_id:
        description: 宿舍ID
        type: string
      user_id:
        description: 用户ID
        type: string
    required:
    - room_id
    - user_id
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormCheckOutReq:
    properties:
      room_id:
        description: 宿舍ID
        type: string
      user_id:
        description: 用户ID
        type: string
    required:
    - room_id
    - user_id
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormCreateReq:
    properties:
      building_id:
        description: 楼栋ID
        type: string
      floor:
        description: 楼层
        type: integer
      max_count:
        description: 最大入住人数
        minimum: 1
        type: integer
      remark:
        description: 备注
        maxLength: 500
        type: string
      room_no:
        description: 宿舍编号
        maxLength: 50
        type: string
      room_type:
        description: 宿舍类型：1-4人间 2-6人间 3-8人间 4-其他
        enum:
        - 1
        - 2
        - 3
        - 4
        type: integer
    required:
    - building_id
    - floor
    - max_count
    - room_no
    - room_type
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormListPageReq:
    properties:
      building_id:
        description: 楼栋ID
        type: string
      campus_id:
        description: 校区ID
        type: string
      currentPage:
        description: 当前页
        type: integer
      floor:
        description: 楼层
        type: integer
      pageSize:
        description: 每页条数
        type: integer
      room_type:
        description: 宿舍类型
        type: integer
      status:
        description: 状态：1-可入住 2-已满 3-装修中
        type: integer
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormResult:
    properties:
      available_bed:
        description: 剩余床位
        type: integer
      building_id:
        type: string
      building_name:
        description: 楼栋名称
        type: string
      campus_id:
        type: string
      campus_name:
        description: 校区名称
        type: string
      created_at:
        type: string
      current_count:
        type: integer
      fill_rate:
        description: 入住率
        type: number
      floor:
        type: integer
      id:
        type: string
      max_count:
        type: integer
      remark:
        type: string
      room_no:
        type: string
      room_type:
        description: 宿舍类型：1-4人间 2-6人间 3-8人间 4-其他
        type: integer
      room_type_name:
        type: string
      status:
        description: 状态：1-可入住 2-已满 3-装修中
        type: integer
      updated_at:
        type: string
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormTransferReq:
    properties:
      room_id:
        description: 当前宿舍ID
        type: string
      target_room_id:
        description: 目标宿舍ID
        type: string
      user_id:
        description: 用户ID
        type: string
    required:
    - room_id
    - target_room_id
    - user_id
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormUpdateReq:
    properties:
      building_id:
        description: 楼栋ID
        type: string
      floor:
        description: 楼层
        type: integer
      id:
        description: 宿舍ID
        type: string
      max_count:
        description: 最大入住人数
        minimum: 1
        type: integer
      remark:
        description: 备注
        maxLength: 500
        type: string
      room_no:
        description: 宿舍编号
        maxLength: 50
        type: string
      room_type:
        description: 宿舍类型：1-4人间 2-6人间 3-8人间 4-其他
        enum:
        - 1
        - 2
        - 3
        - 4
        type: integer
    required:
    - building_id
    - floor
    - id
    - max_count
    - room_no
    - room_type
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormUserListReq:
    properties:
      currentPage:
        description: 当前页
        type: integer
      pageSize:
        description: 每页条数
        type: integer
      room_id:
        description: 宿舍ID
        type: string
      status:
        description: 状态：1-在住 2-已迁出
        type: integer
      user_id:
        description: 用户学号/工号
        type: string
      user_name:
        description: 用户姓名
        type: string
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormUserResult:
    properties:
      building_name:
        description: 楼栋名称
        type: string
      campus_name:
        description: 校区名称
        type: string
      check_in_time:
        type: string
      check_out_time:
        type: string
      id:
        type: string
      room_id:
        type: string
      room_no:
        description: 宿舍编号
        type: string
      status:
        description: 状态：1-在住 2-已迁出
        type: integer
      user_id:
        type: string
      user_name:
        type: string
      user_type:
        type: string
    type: object
...
  github_com_Struggle-Rabbit_CampusLogistics_api_dto.PageResult:
    properties:
      currentPage:
        description: 当前页
        type: integer
      list:
        description: 数据列表
      pageSize:
        description: 每页条数
        type: integer
      total:
        description: 总条数
        type: integer
    type: object
...
  github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse:
    properties:
      code:
        example: 400
        type: integer
      err:
        type: object
      msg:
        example: 失败
        type: string
    type: object
  github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      data:
        type: object
      msg:
        example: 操作成功
        type: string
    type: object
host: localhost:8880
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for user management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /api/v1/dorm/assign:
    post:
      consumes:
      - application/json
      description: 为学生分配宿舍，会检查床位是否已满、用户是否已在其他宿舍入住
      parameters:
      - description: 分配参数，room_id为宿舍ID，user_id为用户ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormAssignReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 分配宿舍接口
      tags:
      - 宿舍管理
  /api/v1/dorm/checkout:
    post:
      consumes:
      - application/json
      description: 办理学生退宿手续，自动释放床位
      parameters:
      - description: 退宿参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormCheckOutReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 退宿接口
      tags:
      - 宿舍管理
  /api/v1/dorm/create:
    post:
      consumes:
      - application/json
      description: 创建新的宿舍信息，包含宿舍编号、所属楼栋、楼层、房间类型、床位数等
      parameters:
      - description: 宿舍创建请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 创建宿舍接口
      tags:
      - 宿舍管理
  /api/v1/dorm/del:
    post:
      consumes:
      - application/json
      description: 根据宿舍ID列表删除宿舍信息，删除前会检查是否有学生入住
      parameters:
      - description: 删除参数，ids为宿舍ID数组
        in: body
        name: data
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 删除宿舍接口
      tags:
      - 宿舍管理
  /api/v1/dorm/detail:
    get:
      consumes:
      - application/json
      description: 根据宿舍ID获取宿舍的详细信息，包含楼栋名称、校区名称、床位使用情况
      parameters:
      - description: 宿舍ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 获取宿舍详情接口
      tags:
      - 宿舍管理
  /api/v1/dorm/list:
    get:
      consumes:
      - application/json
      description: 分页查询宿舍信息列表，支持按楼栋、校区、楼层、宿舍类型、状态筛选
      parameters:
      - description: 查询参数
        in: body
        name: data
        schema:
          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormListPageReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormResult'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 分页查询宿舍列表接口
      tags:
      - 宿舍管理
  /api/v1/dorm/transfer:
    post:
      consumes:
      - application/json
      description: 将学生从当前宿舍调整到目标宿舍，会检查目标宿舍是否有空床位
      parameters:
      - description: 调宿参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormTransferReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 调宿接口
      tags:
      - 宿舍管理
  /api/v1/dorm/update:
    post:
      consumes:
      - application/json
      description: 根据宿舍ID更新宿舍信息，包括宿舍编号、楼层、房间类型、床位数等
      parameters:
      - description: 宿舍更新请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 更新宿舍接口
      tags:
      - 宿舍管理
  /api/v1/dorm/users:
    get:
      consumes:
      - application/json
      description: 获取指定宿舍的所有在住/历史入住人员列表
      parameters:
      - description: 查询参数
        in: body
        name: data
        schema:
          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormUserListReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormUserResult'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 获取宿舍人员列表接口
      tags:
      - 宿舍管理
  /api/v1/dorm/warning:
    get:
      consumes:
      - application/json
      description: 获取入住率达到80%及以上的宿舍列表，用于预警提醒
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_api_dto.DormResult'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Struggle-Rabbit_CampusLogistics_pkg_utils.ErrResponse'
      summary: 获取宿舍容量预警接口
      tags:
      - 宿舍管理
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"